Updated Program
I have made the classes unnested, so I have separated the classes. I have added a method to ProgramSemester as we need to separate classes based on which semester we are in.

Please see, ProgramWithModule, ProgramSemester, ProgramYear.

I have tested in the main, but removed the code and kept like this, so that I can commit it after we've all had a look.